move dialer control to outline package
