--- client/go/outline/vpn/dialer_linux.go
@@ -1,39 +1,0 @@
-// Copyright 2024 The Outline Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package vpn
-
-import (
-	"errors"
-	"syscall"
-)
-
-// TCPDialerControl returns a ControlFn that sets the SO_MARK socket option on a TCP connection.
-// This is used to exclude traffic targeting the remote proxy server from routing to TUN device.
-func TCPDialerControl(conf *Config) (ControlFn, error) {
-	if conf == nil {
-		return nil, errors.New("VPN config must be provided")
-	}
-	return func(network, address string, c syscall.RawConn) error {
-		return c.Control(func(fd uintptr) {
-			syscall.SetsockoptInt(int(fd), syscall.SOL_SOCKET, syscall.SO_MARK, int(conf.ProtectionMark))
-		})
-	}, nil
-}
-
-// TCPDialerControl returns a ControlFn that sets the SO_MARK socket option on a TCP connection.
-// This is used to exclude traffic targeting the remote proxy server from routing to TUN device.
-func UDPDialerControl(conf *Config) (ControlFn, error) {
-	return TCPDialerControl(conf)
-}
